skillLevel separately
